Refactor part 3
Summary:

Test Plan:
